atlas_v6/requirements.txt
discord.py==2.4.0
aiohttp==3.9.5
feedparser==6.0.11
pytz==2024.1
